requirements.txt
tensorflow-gpu
keras
scikit-image
numpy
tqdm
opencv-python
